# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: مصنع نبع صافيا - نظام إدارة المخزن والمبيعات
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reps:
    get:
      operationId: listReps
      tags: [reps]
      summary: قائمة المندوبين
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rep"
    post:
      operationId: createRep
      tags: [reps]
      summary: إضافة مندوب
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rep"
  /reps/{id}:
    get:
      operationId: getRep
      tags: [reps]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rep"
    patch:
      operationId: updateRep
      tags: [reps]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rep"
    delete:
      operationId: deleteRep
      tags: [reps]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ===================== PURCHASES =====================
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: قائمة طلبيات البيع
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: repId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: إضافة طلبية بيع (تخصم من المخزن تلقائياً)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    delete:
      operationId: deleteSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ===================== INVENTORY =====================
...
  /treasury/transactions:
    get:
      operationId: listTreasuryTransactions
      tags: [treasury]
      summary: حركات الخزينة
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [in, out]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TreasuryTransaction"
    post:
      operationId: createTreasuryTransaction
      tags: [treasury]
      summary: إضافة حركة للخزينة
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreasuryTransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreasuryTransaction"

  # ===================== REPORTS =====================
...
components:
  schemas:
...
    Rep:
      type: object
      required: [id, name, phone, area, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        totalSales:
          type: number
        totalCollections:
          type: number
    RepInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        area:
          type: string
    RepUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        area:
          type: string
        isActive:
          type: boolean

    # ---- PURCHASE ----
...
    Sale:
      type: object
      required: [id, orderNumber, date, customerId, customerName, repId, repName, items, totalAmount, vatAmount, grandTotal, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        date:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        repId:
          type: integer
        repName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        totalAmount:
          type: number
        vatAmount:
          type: number
        grandTotal:
          type: number
        paymentMethod:
          type: string
          enum: [coupon, cash, network]
        couponBookId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SaleItem:
      type: object
      required: [productId, productName, quantity, unitPrice, vatRate, subtotal]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        vatRate:
          type: number
        subtotal:
          type: number
    SaleInput:
      type: object
      required: [date, customerId, repId, items, paymentMethod]
      properties:
        orderNumber:
          type: string
        date:
          type: string
        customerId:
          type: integer
        repId:
          type: integer
        paymentMethod:
          type: string
          enum: [coupon, cash, network]
        couponBookId:
          type: integer
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
    SaleItemInput:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
        vatRate:
          type: number

    # ---- INVENTORY ----
...
    TreasuryTransaction:
      type: object
      required: [id, date, type, amount, description, source, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
        type:
          type: string
          enum: [in, out]
        amount:
          type: number
        description:
          type: string
        source:
          type: string
          enum: [sale, expense, purchase, manual]
        reference:
          type: ["string", "null"]
        createdAt:
          type: string
    TreasuryTransactionInput:
      type: object
      required: [date, type, amount, description]
      properties:
        date:
          type: string
        type:
          type: string
          enum: [in, out]
        amount:
          type: number
        description:
          type: string
        notes:
          type: string

    # ---- REPORTS ----
...